requirements.txt
Flask>=2.0.0
google-cloud-bigquery>=2.0.0
gunicorn==20.1.0
pandas
db-dtypes>=1.0.0